Remove code from RobolectricGradleTestRunner since this functionality has been implemented in the ManifestFactories. Add missing abi split functionality that got dropped. Deprecate RobolectricGradleTestRunner in favor of using RobolectricTestRunner (#2517)
